backend/requirements.txt
Flask
flask-cors
tensorflow
opencv-python
Pillow
gdown
gunicorn